split regex and norm
