+++ include/minishell.h
@@ -6,7 +6,7 @@
 /*   By: gakarbou <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/22 21:07:29 by gakarbou          #+#    #+#             */
-/*   Updated: 2025/03/27 03:33:41 by lroussel         ###   ########.fr       */
+/*   Updated: 2025/03/27 05:13:41 by lroussel         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -135,6 +135,6 @@ void			unset_var(char *name, t_list **envp, t_main_envp *imp);
 
 void			init_signals(void);
 
-t_list	*process_wildcards(char *path, char *pattern);
+t_list			*process_wildcards(char *path, char *pattern);
 
 #endif
